# PERF: cache _get_cython_function in groupby ops